requirements.txt
fastapi==0.115.0
firebase-admin==6.6.0
email-validator==2.2.0
google-cloud-firestore>=2.19.0,<3.0
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.20.2
google-cloud-aiplatform==1.71.1
python-dotenv==1.0.1
uvicorn[standard]==0.30.6

